Move classes around to difference folders for organization.
